Totally refactor routines to use significantly less memory
Instead of storing all the tasks in memory, I now only store the routine
name and the indexes to the tasks in memory while leaving the task info
in the store. One buffer of 100 len string is used to display the
current task on the screen and that same buffer is re-used.